linkcache: add ability to have configurable line rewriters
We have @google handled as a special case, but there's no reason it
should be. This commit adds a new rewriter feature that allows the user
to embed python code in their config file along with a regex that
matches a line to be rewritten.

Usage:
[general]
rewriters: rewriter_name

[rewriter_name]
regex: ^(.*)$
rewriter: line = "http://www.google.com/%s" % match.group(1)

The rewriter is passed one argument, 'match', that is the regex match. It
is never None. The new line to be used is assigned to 'line'. None will be
ignored.

Signed-off-by: Jeff Mahoney <jeffm@jeffreymahoney.com>