wolframalphahelper: add did-you-mean, and best-effort value picking
For some queries that would normally return no results, WA offers
a did-you-mean option that suggests a replacement query. We can present
that via a new AmbiguousResult exception. We can also just use the first
and second pods to take a result from if none of the primary ones
are offered.

Signed-off-by: Jeff Mahoney <jeffm@jeffreymahoney.com>